+++ e3nn/experimental/__init__.py
@@ -1,6 +1,5 @@
 from ._full_tp import FullTensorProduct as FullTensorProductv2
 from ._linear import Linearv2
-from ._tp import TensorProduct
 from ._basic import from_chunks
 
-__all__ = [FullTensorProductv2, Linearv2, TensorProduct, from_chunks]
+__all__ = [FullTensorProductv2, Linearv2, from_chunks]
